solve error
