+++ backend/src/modules/restaurant/routes/showDetail.ts
@@ -3,6 +3,7 @@ import { getRestaurant } from ".."
 import { Restaurant } from "@apiType/restaurant"
 const showDetail = (req: Request, res: Response) => {
     const id = parseInt(req.params.id)
+    console.log(id)
     let selectedRes: Restaurant | null = null
     getRestaurant().forEach((res) => {
         if (res.id == id) {
