--- frontend/src/pages/restaurant/data/detailHistory.ts
@@ -1,16 +1,0 @@
-export const detailHistory = [
-    {
-        resName: "Kitchen cheif's",
-        date: "12-11-2022",
-        status: true
-    {
-        resName: "Kitchen cheif's",
-        date: "13-11-2022",
-        status: false
-    },
-    {
-        resName: "Kitchen cheif's",
-        date: "14-11-2022",
-        status: true
-    }
-]
