--- frontend/src/pages/restaurant/data/imgRes.ts
@@ -1,17 +1,0 @@
-export const imgRes = [
-    {
-        img1: "https://upload.wikimedia.org/wikipedia/commons/thumb/1/1e/Tom%27s_Restaurant%2C_NYC.jpg/800px-Tom%27s_Restaurant%2C_NYC.jpg?20170523012006",
-        img2: "https://img.freepik.com/free-vector/hand-painted-watercolor-pastel-sky-background_23-2148902771.jpg?w=2000",
-        img3: "https://img.freepik.com/free-photo/grunge-paint-background_1409-1337.jpg?w=2000",
-    },
-    {
-        img1: "https://img.freepik.com/free-photo/grunge-paint-background_1409-1337.jpg?w=2000",
-        img2: "https://img.freepik.com/free-vector/hand-painted-watercolor-pastel-sky-background_23-2148902771.jpg?w=2000",
-        img3: "https://upload.wikimedia.org/wikipedia/commons/thumb/1/1e/Tom%27s_Restaurant%2C_NYC.jpg/800px-Tom%27s_Restaurant%2C_NYC.jpg?20170523012006",
-    },
-    {
-        img1: "https://img.freepik.com/free-vector/hand-painted-watercolor-pastel-sky-background_23-2148902771.jpg?w=2000",
-        img2: "https://img.freepik.com/free-photo/grunge-paint-background_1409-1337.jpg?w=2000",
-        img3: "https://img.freepik.com/free-photo/grunge-paint-background_1409-1337.jpg?w=2000",
-    },
-]
