--- frontend/src/pages/restaurant/data/review.ts
@@ -1,128 +1,0 @@
-export const Review = [
-    {
-        resId: 0,
-        name: "Elon Malabo",
-        picture: "",
-        rate: "4",
-        review: "This restaurant is so good but it is a little pricey will come back ",
-    },
-    {
-        resId: 1,
-        name: "Elon Malabo",
-        picture: "",
-        rate: "5",
-        review: "This restaurant is so good",
-    },
-    {
-        resId: 0,
-        name: "MarK Tole",
-        picture: "",
-        rate: "5",
-        review: "This restaurant is the best",
-    },
-    {
-        resId: 0,
-        name: "John Farmer",
-        picture: "",
-        rate: "3",
-        review: "This restaurant pricey ",
-    },
-    {
-        resId: 0,
-        name: "El Ham",
-        picture: "",
-        rate: "4",
-        review: "This restaurant is so good but it is a little pricey will come back ",
-    },
-    {
-        resId: 1,
-        name: "Leo Dosth",
-        picture: "",
-        rate: "5",
-        review: "This restaurant is vary nice",
-    },
-    {
-        resId: 1,
-        name: "Wale Gio",
-        picture: "",
-        rate: "5",
-        review: "tasty ",
-    },
-    {
-        resId: 2,
-        name: "Tom Fol",
-        picture: "",
-        rate: "4",
-        review: "Good",
-    },
-    {
-        resId: 2,
-        name: "Roe Tui",
-        picture: "",
-        rate: "4",
-        review: "Ahhhhh so muchh",
-    },
-    {
-        resId: 3,
-        name: "Suki Oni",
-        picture: "",
-        rate: "5",
-        review: "soo goodddddd ",
-    },
-    {
-        resId: 4,
-        name: "Kily fox",
-        picture: "",
-        rate: "0",
-        review: "who liked this restaurant ",
-    },
-    {
-        resId: 4,
-        name: "Suki Oni",
-        picture: "",
-        rate: "5",
-        review: "soo goodddddd ",
-    },
-    {
-        resId: 5,
-        name: "Kily fox",
-        picture: "",
-        rate: "0",
-        review: "who liked this restaurant ",
-    },
-    {
-        resId: 6,
-        name: "Roe Tui",
-        picture: "",
-        rate: "4",
-        review: "Ahhhhh so muchh",
-    },
-    {
-        resId: 7,
-        name: "Suki Oni",
-        picture: "",
-        rate: "5",
-        review: "soo goodddddd ",
-    },
-    {
-        resId: 7,
-        name: "Kily fox",
-        picture: "",
-        rate: "0",
-        review: "who liked this restaurant ",
-    },
-    {
-        resId: 8,
-        name: "Suki Oni",
-        picture: "",
-        rate: "5",
-        review: "soo goodddddd ",
-    },
-    {
-        resId: 9,
-        name: "Kily fox",
-        picture: "",
-        rate: "0",
-        review: "who liked this restaurant ",
-    },
-]
