+++ frontend/src/pages/restaurant/detail/[detailRes].tsx
